ElementBase abstract base class
