+++ lib/kitchen.rb
@@ -16,6 +16,7 @@ end
 
 require_all("kitchen/mixins")
 
+require "kitchen/utils"
 require "kitchen/errors"
 require "kitchen/ancestor"
 require "kitchen/element_enumerator"
@@ -27,16 +28,8 @@ require "kitchen/book_element_enumerator"
 require "kitchen/figure_element_enumerator"
 require "kitchen/table_element_enumerator"
 require "kitchen/example_element_enumerator"
-require "kitchen/document"
+require "kitchen/element_base"
 require "kitchen/element"
-require "kitchen/debug/print_recipe_error"
-require "kitchen/recipe"
-require "kitchen/book_recipe"
-require "kitchen/oven"
-require "kitchen/clipboard"
-require "kitchen/pantry"
-require "kitchen/counter"
-require "kitchen/book_document"
 require "kitchen/book_element"
 require "kitchen/chapter_element"
 require "kitchen/example_element"
@@ -45,6 +38,15 @@ require "kitchen/term_element"
 require "kitchen/figure_element"
 require "kitchen/table_element"
 require "kitchen/note_element"
+require "kitchen/document"
+require "kitchen/book_document"
+require "kitchen/debug/print_recipe_error"
+require "kitchen/recipe"
+require "kitchen/book_recipe"
+require "kitchen/oven"
+require "kitchen/clipboard"
+require "kitchen/pantry"
+require "kitchen/counter"
 require "kitchen/note_element_enumerator"
 require "kitchen/patches"
